Still store session data and send cookie if other middlewares throw
If a middleware throws an error, and that error does not get caught
before entering back into the session middleware, cookies were not sent
and session data was not stored.

This can happen if users have a 'problem' (error handling) middleware
before the session middleware. Errors bubble right up.

Using try...finally we can ensure that the session storing mechanism
always worked, regardless of if `await next()` threw an error or not.

The included test-case failed before this change, but passes after this
change.